review changes: moved button outside of dialog
