+++ services/frontend-service/src/ui/Warnings.tsx
@@ -1,7 +1,24 @@
+/*This file is part of kuberpult.
+
+Kuberpult is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Kuberpult is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with kuberpult.  If not, see <http://www.gnu.org/licenses/>.
+
+Copyright 2021 freiheit.com*/
 import * as React from 'react';
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
 import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
 import type { Environment, Release } from '../api/api';
+import { Tooltip } from '@material-ui/core';
 
 export enum DeployState {
     Normal,
@@ -34,8 +51,10 @@ export function getDeployState(name: string, environments: { [name: string]: Env
 const UndeployRunningWarning: React.FC<any> = (props: { deployState: DeployState }) => {
     const tooltip = 'This app is in the process of deletion';
     const undeployHint = (
-        <div className={'warning-undeploy-running'} title={tooltip}>
-            <DeleteOutlineIcon />
+        <div className={'warning-undeploy-running'}>
+            <Tooltip title={tooltip}>
+                <DeleteOutlineIcon />
+            </Tooltip>
         </div>
     );
     switch (props.deployState) {
